Tests for values ranges

+++ src/values-range.h
@@ -8,4 +8,4 @@ double get_theta_I_half_count(double d_e, double d_m);
 
 int get_N_theta(double d_e, double d_m, int N_b);
 
-int get_h_max(double d_e, double d_m, int N_b);
+double get_h_max(double d_e, double d_m, int N_b);
